Support single agent binary on Windows
