+++ cmd/agent/bundle.go
@@ -3,21 +3,17 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build bundle_trace_agent
-
 // Main package for the agent binary
 package main
 
 import (
-	"os"
-
-	tracecommand "github.com/DataDog/datadog-agent/cmd/trace-agent/command"
 	"github.com/spf13/cobra"
 )
 
-func init() {
-	registerAgent([]string{"trace-agent"}, func() *cobra.Command {
-		os.Args = tracecommand.FixDeprecatedFlags(os.Args, os.Stdout)
-		return tracecommand.MakeRootCommand()
-	})
+var agents = map[string]func() *cobra.Command{}
+
+func registerAgent(names []string, getCommand func() *cobra.Command) {
+	for _, name := range names {
+		agents[name] = getCommand
+	}
 }
