+++ cmd/agent/process_agent.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build !windows && bundle_process_agent
+//go:build bundle_process_agent
 
 // Main package for the agent binary
 package main
