+++ cmd/agent/system_probe_windows.go
@@ -9,13 +9,30 @@
 package main
 
 import (
+	"os"
+
+	"github.com/DataDog/datadog-agent/cmd/security-agent/windows/service"
 	sysprobecommand "github.com/DataDog/datadog-agent/cmd/system-probe/command"
 	sysprobesubcommands "github.com/DataDog/datadog-agent/cmd/system-probe/subcommands"
+	"github.com/DataDog/datadog-agent/pkg/util/log"
+	"github.com/DataDog/datadog-agent/pkg/util/winutil/servicemain"
+
 	"github.com/spf13/cobra"
 )
 
 func init() {
 	registerAgent([]string{"system-probe"}, func() *cobra.Command {
+		// if command line arguments are supplied, even in a non-interactive session,
+		// then just execute that.  Used when the service is executing the executable,
+		// for instance to trigger a restart.
+		if len(os.Args) == 1 {
+			if servicemain.RunningAsWindowsService() {
+				servicemain.Run(&service.Service{})
+				return nil
+			}
+		}
+		defer log.Flush()
+
 		rootCmd := sysprobecommand.MakeCommand(sysprobesubcommands.SysprobeSubcommands())
 		sysprobecommand.SetDefaultCommandIfNonePresent(rootCmd)
 		return rootCmd
