+++ cmd/system-probe/main_windows.go
@@ -9,52 +9,15 @@
 package main
 
 import (
-	"context"
-	"errors"
-	"fmt"
 	"os"
 
 	"github.com/DataDog/datadog-agent/cmd/internal/runcmd"
 	"github.com/DataDog/datadog-agent/cmd/system-probe/command"
-	"github.com/DataDog/datadog-agent/cmd/system-probe/config"
 	"github.com/DataDog/datadog-agent/cmd/system-probe/subcommands"
-	runsubcmd "github.com/DataDog/datadog-agent/cmd/system-probe/subcommands/run"
+	"github.com/DataDog/datadog-agent/cmd/system-probe/windows/service"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/util/winutil/servicemain"
 )
-
-type service struct {
-	servicemain.DefaultSettings
-	errChan <-chan error
-	ctxChan chan context.Context
-}
-
-func (s *service) Name() string {
-	return config.ServiceName
-}
-
-func (s *service) Init() error {
-	s.ctxChan = make(chan context.Context)
-
-	errChan, err := runsubcmd.StartSystemProbeWithDefaults(s.ctxChan)
-	if err != nil {
-		if errors.Is(err, runsubcmd.ErrNotEnabled) {
-			return fmt.Errorf("%w: %w", servicemain.ErrCleanStopAfterInit, err)
-		}
-		return err
-	}
-
-	s.errChan = errChan
-
-	return nil
-}
-
-func (s *service) Run(ctx context.Context) error {
-	// send context to background agent goroutine so we can stop the agent
-	s.ctxChan <- ctx
-	// wait for agent to stop
-	return <-s.errChan
-}
 
 func main() {
 	// if command line arguments are supplied, even in a non-interactive session,
@@ -62,7 +25,7 @@ func main() {
 	// for instance to trigger a restart.
 	if len(os.Args) == 1 {
 		if servicemain.RunningAsWindowsService() {
-			servicemain.Run(&service{})
+			servicemain.Run(&service.Service{})
 			return
 		}
 	}
