+++ tasks/security_agent.py
@@ -64,7 +64,7 @@ def build(
     """
     Build the security agent
     """
-    if bundle and sys.platform != "win32":
+    if bundle:
         return agent_build(
             ctx,
             install_path=install_path,
